move patient login into lib
