+++ lib/index.ts
@@ -1,11 +1,12 @@
-import Crypto from "crypto"
-import jwt from "jsonwebtoken"
-import Path from "path"
-import { NextFunction, Request, Response, RequestHandler } from "express"
-import { readdirSync, statSync } from "fs"
-import { HttpError, InvalidRequestError, OAuthError } from "./errors"
-import config from "../config"
-import { EHI } from "../index"
+import Crypto                                               from "crypto"
+import jwt                                                  from "jsonwebtoken"
+import Path                                                 from "path"
+import { NextFunction, Request, Response, RequestHandler }  from "express"
+import { readdirSync, statSync }                            from "fs"
+import { HttpError, InvalidRequestError, OAuthError }       from "./errors"
+import patients                                             from "../data/db"
+import config                                               from "../config"
+import { EHI }                                              from "../index"
 
 
 /**
@@ -286,6 +287,29 @@ export async function logout(req: EHI.UserRequest, res: Response) {
     await wait(config.authDelay);
     delete req.user!.sid;
     return res.clearCookie("sid").end("Logout successful");
+}
+
+export function patientLogin(req: EHI.UserRequest, res: Response) { 
+    const list: any[] = [];
+    patients.forEach((value, key) => {
+        list.push({ id: key, name: value.patient.name, birthDate: value.patient.birthDate })
+    })
+    
+    // Turn some unique visitor information (e.g. IP) into a patient-index to
+    // promote to the front of the list, reducing patient-collisions across multiple users
+    if (list.length > 0) { 
+        const seed = req.ip;
+        const hash = Crypto.createHash('sha256'); 
+        hash.update(seed)
+        const hexValue = hash.digest('hex')
+        // Use last ten digits only to avoid generating really large numbers.
+        // We might lose trailing-digit precision when dealing with massive floats 
+        const uniqueValue = parseInt(hexValue.slice(hexValue.length - 10, hexValue.length), 16);
+        const indexToPromote = uniqueValue % patients.size;
+        [list[0], list[indexToPromote]] = [list[indexToPromote], list[0]]
+    }
+    
+    res.render("patient-login", { patients: list, query: req.query })
 }
 
 export async function waitFor(condition: () => any): Promise<void> {
